requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
apscheduler==3.10.4
slack-sdk==3.23.0
python-dotenv==1.0.0
webdriver-manager==4.0.1
plotly==5.17.0
pandas==2.1.3
pytz==2023.3
